+++ docs/changes.rst
@@ -5,6 +5,8 @@ Weblate 5.17
 
 .. rubric:: New features
 
+* Added :setting:`WEBSITE_ALERTS_ENABLED` setting to allow disabling project website availability checks and alerts.
+
 .. rubric:: Improvements
 
 * Track origin of newly added source strings.